refactor: move light-client APIs to own crate, migrate sha impl
