+++ packages/light-client/README.md
@@ -1,0 +1,8 @@
+# @near-js/light-client
+
+NEAR light client verification utilities. Based on the [Light Client spec](https://github.com/near/NEPs/blob/master/specs/ChainSpec/LightClient.md)
+
+# License
+
+This repository is distributed under the terms of both the MIT license and the Apache License (Version 2.0).
+See [LICENSE](https://github.com/near/near-api-js/blob/master/LICENSE) and [LICENSE-APACHE](https://github.com/near/near-api-js/blob/master/LICENSE-APACHE) for details.
