+++ packages/light-client/tsconfig.json
@@ -1,0 +1,9 @@
+{
+    "extends": "../../tsconfig.node.json",
+    "compilerOptions": {
+        "outDir": "./lib",
+    },
+    "files": [
+        "src/index.ts"
+    ]
+}
