+++ packages/near-api-js/package.json
@@ -22,6 +22,7 @@
         "@near-js/types": "workspace:*",
         "@near-js/utils": "workspace:*",
         "@near-js/wallet-account": "workspace:*",
+        "@near-js/light-client": "workspace:*",
         "ajv": "^8.11.2",
         "ajv-formats": "^2.1.1",
         "bn.js": "5.2.1",
